adding calcular ganancias

+++ src/Project/models/CultivoSeleccionadoV2.java
@@ -58,7 +58,7 @@ public class CultivoSeleccionadoV2 {
         return (int) this.riesgoAsociado;
     }
 
-    public void setRiesgoAsociado(int riesgoAsociado) {
+    public void setRiesgoAsociado(double riesgoAsociado) {
         this.riesgoAsociado = riesgoAsociado;
     }
